+++ Demo/ViewModels/UnitsViewModel.cs
@@ -4,55 +4,28 @@ using DigitalProduction.ViewModels;
 using System.Collections.ObjectModel;
 using Thor.Units;
 
-namespace Data.Translation.ViewModels;
+namespace UnitsConversionDemo.ViewModels;
 
-public abstract partial class UnitsViewModel : DataGridBaseViewModel<UnitGroup>
+public partial class UnitsViewModel : UnitsViewModelBase
 {
 
 	#region Fields
-
-	[ObservableProperty]
-	private UnitConverter?						_unitsConverter;
 
 	#endregion
 
 	public UnitsViewModel()
     {
-		OnUnitsChanged();
+		UnitsConverter = UnitFileIO.LoadVersionTwoFile();
 	}
 
-
-	private void OnUnitsChanged()
-	{
-		try
-		{
-			UnitGroup[]? unitGroups = UnitsConverter?.Groups.GetAllGroups();
-			Items = unitGroups != null ? new ObservableCollection<UnitGroup>(unitGroups) : null;
-		}
-		catch
-		{
-			Items = null;
-		}
-	}
-
-	[RelayCommand]
-	private async Task<bool> SaveUnits()
-	{
-		if (UnitsConverter != null)
-		{
-			bool success = await SerializeAsync();
-			if (success)
-			{
-				Modified = false;
-			}
-			return success;
-		}
-
-		return false;
-	}
 
 	/// <summary>
 	/// Write this object to a file.  The Path must be set and represent a valid path or this method will throw an exception.
 	/// </summary>
-	public abstract Task<bool> SerializeAsync();
+	async public override Task<bool> SerializeAsync()
+	{
+		//return true;
+		return true;
+
+	}
 }
